added tasks

+++ week-01/day-04/Parametric/src/Parametric.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Parametric {
     public static void main(String[] args) {
 
-        double a = 0;
+        int a = 0;
         System.out.println("Give me a number, please. ");
         Scanner scan = new Scanner(System.in);
 
@@ -12,7 +12,7 @@ public class Parametric {
 
         for (int i = 0; i < numb; i++) ;
         int b = scan.nextInt();
-        double a = (a + b);
+        int a = (a + b);
 
         System.out.println(a / numb + " is the average");
         System.out.println(a + " is the amount");
